% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromePlus-package.R
\docType{package}
\name{chromePlus-package}
\alias{chromePlus}
\alias{chromePlus-package}
\title{chromePlus: Analysis of Chromosome Number Evolution and Binary Traits}
\description{
Implements models of chromosome number evolution similar to chromEvol using Markov and MuSSE models in the diversitree framework. Supports state-dependent chromosome evolution where a binary trait can affect rates of chromosome gain, loss, polyploidy, and demiploidy, as well as state-dependent speciation and extinction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coleoguy/chromePlus}
  \item Report bugs at \url{https://github.com/coleoguy/chromePlus/issues}
}

}
\author{
\strong{Maintainer}: Heath Blackmon \email{coleoguy@gmail.com} (\href{https://orcid.org/0000-0002-5433-4036}{ORCID})

Authors:
\itemize{
  \item Michelle M. Jonika (\href{https://orcid.org/0000-0002-1170-6418}{ORCID})
  \item Emma Goldberg (\href{https://orcid.org/0000-0002-7311-9606}{ORCID})
  \item Itay Mayrose (\href{https://orcid.org/0000-0002-8460-1502}{ORCID})
}

Other contributors:
\itemize{
  \item Maximos Chin [contributor]
}

}
\keyword{internal}
