\name{chomevol.format.txt}
\alias{params.txt}
\title{
Data and control file formatted for the program chromEvol
}
\description{
These files can be used with the included tree file to run a comparison between chromePlus and chromEvol}
\usage{
system.file("extdata", "chomevol.format.txt", package = "chromevolR")
system.file("extdata", "params.txt", package = "chromevolR")
}

\references{
http://www.tau.ac.il/~itaymay/cp/chromEvol/
}
\author{
Heath Blackmon}

\examples{


}
